+++ NEWS.md
@@ -2,7 +2,7 @@ Development - YYYY-MM-DD
 --------------------------
 New Features:
 
-* New function adjust.hsv() to adjust colors in HSV space.
+* New function adjust_hsv() to adjust colors in HSV space.
 
 Release 3.2.0 - 2024-09-28
 --------------------------